fix(dev): move queries out of /api/ to avoid vite proxy collision
In dev, vite proxies /api/* to the FastAPI backend so module
requests like /api/queries.ts were being answered with JSON,
breaking the page. Move client/api/queries.ts to client/lib/
queries.ts and update imports.
